requirements.txt
django
djangorestframework
djangorestframework-simplejwt
dataclasses
django-filter
django-rest-swagger
drf-yasg